// src/Components/redux/redux.types.ts
export enum ActionType {
  SET_CURRENT_SCREEN = "SET_CURRENT_SCREEN",
  SET_LEVEL_NUMBER = "SET_LEVEL_NUMBER",
  SET_SCREENSCALE_FACTOR = "SET_SCREENSCALE_FACTOR",
  SET_MATCHED_ELEMENTS = "SET_MATCHED_ELEMENTS",
  SET_PLAYER1_SCORE = "SET_PLAYER1_SCORE",
  SET_PLAYER2_SCORE = "SET_PLAYER2_SCORE",
  SET_CURRENT_PLAYER = "SET_CURRENT_PLAYER",
  REPLAY_GAME = "REPLAY_GAME",
}